Separate init from constructor
With the addition of subdivision support, which initializes using async
methods, init can be asynchronous. Separate init from the constructor so
that completion of init can be detected. This adds minimal overhead and
only affects init in the event that subdivision support is needed.
Everything else remains sync and fast.